requirements.txt
dm_control>=1.0.12
mujoco>=2.3.5
Pillow
